+++ src/settings/StaffSlips/tokens.js
@@ -118,6 +118,11 @@ const getTokens = (locale) => ({
       allowedFor: allowedForAllStaffSlips,
     },
     {
+      token: 'item.effectiveLocationPrimaryServicePointName',
+      previewValue: 'Circulation Desk',
+      allowedFor: allowedForAllStaffSlips,
+    },
+    {
       token: 'item.effectiveLocationLibrary',
       previewValue: 'Main Library',
       allowedFor: allowedForAllStaffSlips,