+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Add new multiples token: Multiple fee/fine charges for the Fee/fine charge section. Refs UICIRC-904.
 * Remove unnecessary `recordsRequired` manifest param. Refs UIREQ-858.
 * Add "Current date/time" as staff slip token in Settings. Refs UICIRC-793.
+* Add item.effectiveLocationPrimaryServicePointName as staff slip token in Settings. Ref CIRC-1785
 
 ## [8.0.1](https://github.com/folio-org/ui-circulation/tree/v8.0.1) (2023-03-07)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v8.0.0...v8.0.1)